Adds a check for item/mob overlaps
Fixes CoolLord22/OtherDrops#42 With 1.13, several mobs dropped an item
of the same name (for example, chickens now drop chicken rather than
raw_chicken). This conflicts because the plugin defaulted to the entity,
and not the item. Now we check if the user has just put CHICKEN which
resolves to the item, while putting ENTITY_CHICKEN or CREATURE_CHICKEN
will resolve to the mob.

Also updates some material constants