Making it so dropped item can be consumed
By adding consumetool: 1 we can make it so that 1 item that is thrown is consumed. Note that this value cannot be greater than 1, as when the player clicks Q to drop an item only 1 is thrown out. You can instead use a chance parameter to effectively require more of one item to receive the drop. See Wiki for more details!