Fixing mythic toolagent
Fixes a major flaw in mythic tool agents that previously regarded any mythic tool as an "ANY_OBJECT" tool agent; only the override matching method was being used to check if this tool was correct or not. This resulted in a list of tools not being able to trigger the drop properly.

Resolved by making a super call to the parent toolagent class to ensure map contains proper material data.