Update OccurredEvent.java
Fixes a bug that arose when we added in Mythic Mobs as tool agents. The else condition was structured after the dependency check for MythicMobs which prevented the setTool methods from creating creature tools when mythic was present but the mob was not a mythic type.

We now return early when a mythic subject is found, and otherwise set it as the damager entity.

Fixes #53