Adding implementation for MoonPhase
Adds in a new condition for which phase the moon is in on any given Minecraft world day.

Usage is any of the following `"moon", "moons", "moonphase", "moonphases"` how you normally configure conditions.

IMPORTANT NOTE: This only checks for moon phase based on the world's total uptime and performs a math operation to obtain the phase of the moon. This does NOT check for if it is actually night time and that the moon is out. For that, please use the `time` condition in conjunction with `moonphase`.

Accepted value is a list of the following Enum Values `FULL_MOON, WANING_GIBBOUS, LAST_QUARTER, WANING_CRESCENT, NEW_MOON, WAXING_CRESCENT, FIRST_QUARTER, WAXING_GIBBOUS`. See known_lists/MoonPhase.txt

Example:
```yaml
    PLAYER:
        - trigger: HIT
          tool: ANY_CREATURE
          moonphase: [ANY, -FULL_MOON]
          message.server: Not attacked during full moon!
```