Surround namespacedkey in try block
Since we automatically assume an ODItem is getting parsed, when checking for a namespacedkey an IllegalArgumentException would be thrown since it contains characters not supported here. Attempt to convert the parsed key and if its succesful return the ItemStack otherwise return null to indicate this was not a loaded item.