Creates implementation for supporting
offhand.

This implementation creates a new ToolAgent in the PlayerSubject object to track an offhand tool.

Creates issues when the offhand matches, but the mainhand does not, as tooldamage still defaults to main hand.

ver2 will instead un-cancel the hand packet and attempt to process all drops twice, once for main once for off hands.