Refactor of Material/Creature groups
This commit removes the enum structure for MaterialGroup and CreatureGroup classes.

Allows for us to dynamically declare groups on run time (not requiring compile time definition of them).

Allows for Material and Creature groups to be defined using minecraft block/entity tags.

Will open possibility to allow users to define their own groups to be used locally within the configs.

Resolves #60